Created a generic FileDialog UI class used for saving and opening files.
This class takes a number of parameters that alter the way it is displayed
or behaves.

The selection property returns the file selection.  Accessing this property
is what displays the dialog, tests the response, and returns the file
selection if any.

The file manager service uses this new class for displaying dialogs.