Made several improvements to __init__.py and the Gaphor application initialization.
The major change is that Application now has an option parser for access
to command line options.  This means that plugins and services can add command
line options to the parser and have access to their values.  This is done
by creating the parser object inside main() and passing it to the application.
The application will then initialize all services before actually parsing
the options.  This gives them a chance to add their own.

The imports in __init__.py have also been re-arranged.  Logging is now set
when the application is initialized.

The launch() function in __init__.py has been simplified.  Instead of passing
a Gaphor file argument, launch() will now ask the Application instance if
there is a Gaphor model to load.

The main() function in __init__.py is also much smaller and simpler now
due to the above changes.