A few minor changes.
Removed a print statement from application.py

Change the logging format in logger.py.  Instead of showing the asctime in
the message, we now show the logger name.  This isn't a web server so the
time really isn't that relevant.

The Logger constructor now has a name parameter which defaults to GAPHOR and
is displayed in the log messages.  This way new loggers that are created
can have a name and the messages carry more meaning.

The logger class has a new static attribute - level.  This can be set to
apply the logging level globally, for any Logger instances.  For now, this
doesn't have an effect since Logger is a singleton.  However, this is likely
to change.