Improved attribute.load().
I removed conditional that checks if value is an object as it
didn't seem to be relevant based on the generated uml2.py.

Instead, we now try to set the value with a new instance of
self.type directly.  The only way this fails, as far as I
can tell, is by a ValueError when instantiating the type (int('xyz')).

If this fails, we now raise a TypeError with a more descriptive
message.