Fix AttributeError when win action group is None
This fixes two errors caused during creating a new element on a diagram
where it tries to look up an action on the win action group that doesn't
yet exist, and when copying an item caused by the same issue. This fix
adds a guard to check if the win action group is None before looking
up the action.

Signed-off-by: Dan Yeaw <dan@yeaw.me>