Make Placement tool less dependent on diagram toolbox
in the end the tools should end up in their respected diagram folders.
To achieve this I need to untangle the tools with the diagram
toolbox.

The `DiagramItemCreated` event is now emited on item creation.
`DiagramItemPlaced` is a special event emitted from the placement tool
to signal the placement is done.