Fix lifeline ordering when not all items are linked in a diagram
This commit fixes an isue that can occur when loading a model. Lifelines are ordered based on their position on the diagram. The re-ordering is triggered when items are connected. However, relations on model level are already connected at that point. The ordering is dependent on the connections in the diagram as well as the connections on model level.

This change makes the code more resilient during loading.

Fixes #1152