Loading an element should not steal an existing reference
This is particulary important when doing a full (deep) paste of
an item. It could steal element owned by the original, where it
should only copy itself.

Changing this caused the model generation process to change. Elements
are created at different points in time, which caused lines to move
around in the model file. Hence the UML model has been regenerated.