Remove rename entry from popup
On Wayland focus is quickly shifted when the popover closes (I think).
This causes an focus-out signal which causes the edit action to end
instantly.

I can't see a proper way to have the "done-on-focus-out" work in
combination with the popover. Removing the entry from the menu
is a simple solution.