Relax rules for lifeline items
A lifeline item will always support sequence diagram style
messages. If the lifetime (dashed line below the body) is
completely collapsed, the message will show as a communication
diagram message.

In doing so, we allow lines to be connected everywhere on the
lifeline. Making it easier for users to update and change. Also
this allows "create" messages to be attached to the body of a
lifeline item.