Add a priority lane for event handlers
My hypothesis why things still can break is because
not all events are properly registered at the undo manager.
In that case we can no do a full undo, and the model may remain
in an inconsistent state.

To remedy this, we can basically do two things:

1. delay the (model) change until just before the event is about to be
   fired.
2. Make sure all events are captured by the undo manager.

Option 1. is quite complicated, since a simple relation change can
results in a combination of set and delete events.
Option 2. is a bit hacky as well, since we add a "priotity lane"
to the event manager.