Remove redundant translation fallback
When fallback is set to True, gettext uses a NullTranslator to return
the message and an exception isn't raised if there is an issue with the
translation files. We also are using a try-except block to print a
warning message. With fallback set, the warning message would never be
printed.