Support full copy of diagrams
You can now make a full copy of a diagram (from a diagram) and
paste it.

This is done by recording the original diagram id(s). When pasting
the original diagram ids are mapped to the new diagram. The Presentation
paste function does a lookup for the target diagram, which can now also
be a newly created diagram.