Separate item only factory from item with element factory in toolbox and support specifying diagram specific items
In order to support creating diagram specific items for the element,
it is important that element representations are consistent between toolbox added items
and dragged element created items. For that, toolbox no longer is allowed to specify both
item and element - new_element_item() must be used instead.

represents() takes now additional parameter of the diagram class where the item can be represented on.