In postload, only connect the handle
Previously a complete "connect" sequence was performed.
In a proper model all elements are already connected,
so there's no need to perform a connect on model level.
Technically that can even lead to orphaned elements.

Let's *assume* during loading that the model is properly set up.
Anyway, we will only load errors that are also part of the model
at time of saving.