Move diagram update from update_now() to update()
`update_now()` is part if the protocol that Gaphas views use
to communicate with the model (`Diagram`).

By decoupling those we can move away from a model where diagram
updates are done asynchronously.
We need our updates to happen within the context of a transaction.

A possitive side effect may be that tearing, when moveing something in a
diagram, is a thing of the past.