Deal with missed updates due to model changes
For now, just update the diagram as soon as a request for update is
emited. This has the advantage that we can keep the old interface in
tact and we have the option to do smarter things with updates in the
future.

For now, let's just update. We need to make sure that updates happen
before the view is updated.