Cancel transactional drag event sequence on key press
This is a touch one: at all times our transactions must be closed.
When a key is pressed, the event sequence is broken and no
proper finalization happens. Hence, transactions can remain open,
which causes trouble if you want to undo changes.

This commit fixes the case when a key is pressed. There may, however,
be more cases that can cause inconsistency in the transaction stack.