Drop ModelLoaded event
We had three events: SessionCreated, ModelLoaded, and ModelReady.

SessionCreated is emitted when a new session is launched. A session may
be based on an existing model (`filename` is set).

ModelLoaded is emitted when a model is loaded, and it also contains
the file name

ModelReady is emitted when a model is loaded.

There's some overlap in those events. Back in the day we had the option
to load a model in an existing session. We do not allow that anymore.
Therefore the ModelLoaded event is a bit redundant.