Fix creating main window
The window property is used by conditions in multiple places, but it was
always available due to the getter function, which is not the expected
behavior. Turn it into an empty property, and assign the window only in
the open() function, which is the only place where the window is supposed
to be created.

This fixes the problem that set_application() has no effect, because it's
called after the window is presented by _on_file_manager_state_changed().